move: all view funcs into route files; add: login.tsx skeleton
