+++ src/routes/guests.tsx
@@ -1,6 +1,18 @@
-import { createFileRoute } from '@tanstack/react-router'
-import GuestsView from '../views/GuestsView'
+import { createFileRoute } from "@tanstack/react-router";
 
-export const Route = createFileRoute('/guests')({
+export const Route = createFileRoute("/guests")({
   component: GuestsView,
-})
+});
+
+function GuestsView() {
+  return (
+    <>
+      {/*
+      Title: Registered Guests
+      Table (with sort options):
+        fields: UID, First, Last, Has Active Notifications
+      */}
+      <h1>Guests</h1>
+    </>
+  );
+}
