+++ src/routes/new-notification.tsx
@@ -1,6 +1,21 @@
-import { createFileRoute } from '@tanstack/react-router'
-import NewNotificationView from '../views/NewNotificationView'
+import { createFileRoute } from "@tanstack/react-router";
 
-export const Route = createFileRoute('/new-notification')({
+export const Route = createFileRoute("/new-notification")({
   component: NewNotificationView,
-})
+});
+
+function NewNotificationView() {
+  return (
+    <>
+      {/*
+      Title: New Notification
+      Form
+        Guest field
+        Service field
+        Message (optional) field
+      Button: Create Notification - creates notification associated with guest
+    */}
+      <h1>New Notification</h1>
+    </>
+  );
+}
