--- src/views/AddServiceView.tsx
@@ -1,16 +1,0 @@
-import React from 'react'
-
-const AddServiceView = () => {
-  return (<>
-    {/*
-      Title: Add New Service
-      Form
-        Service Name field
-        Service Quota field
-      Button: Save Changes
-    */}
-    <h1>Add Service</h1>
-  </>)
-}
-
-export default AddServiceView
