--- src/views/GuestProfileView.tsx
@@ -1,25 +1,0 @@
-import React from 'react'
-
-const GuestProfileView = () => {
-  return (<>
-    {/*
-      Title: Guest ID
-      Form
-        First name field
-        Last name field
-        Birthday field
-        Case Manager field
-      Button: Save Changes
-
-      Table: Active Notifications
-        fields:
-      Table: Archived Notifications
-        fields:
-      Table: Past Visits
-        fields:
-    */}
-    <h1>Guest Profile</h1>
-  </>)
-}
-
-export default GuestProfileView
