--- src/views/GuestsView.tsx
@@ -1,14 +1,0 @@
-import React from 'react'
-
-const GuestsView = () => {
-  return (<>
-    {/*
-      Title: Registered Guests
-      Table (with sort options):
-        fields: UID, First, Last, Has Active Notifications
-    */}
-    <h1>Guests</h1>
-  </>)
-}
-
-export default GuestsView
