--- src/views/NewNotificationView.tsx
@@ -1,17 +1,0 @@
-import React from 'react'
-
-const NewNotificationView = () => {
-  return (<>
-    {/*
-      Title: New Notification
-      Form
-        Guest field
-        Service field
-        Message (optional) field
-      Button: Create Notification - creates notification associated with guest
-    */}
-    <h1>New Notification</h1>
-  </>)
-}
-
-export default NewNotificationView
