--- src/views/UserProfileView.tsx
@@ -1,17 +1,0 @@
-const UserProfileView = () => {
-  return (<>
-    {/*
-      Title: Username
-      Form
-        Name field
-        Role field
-        email field
-        new password field
-      Button: Save Changes
-      Button (ADMIN): Delete User
-    */}
-    <h1>User Profile</h1>
-  </>)
-}
-
-export default UserProfileView
