--- src/views/VisitsView.tsx
@@ -1,19 +1,0 @@
-import React from 'react'
-
-const VisitsView = () => {
-  return (<>
-    {/*
-      Title: Visits
-      Chart:
-        arrows to click thru months
-        Button: Export Data (CSV)
-        chart with default past month view
-      Table: Queue
-        search / filter
-        fields: time requested, guest name w/ ID, Services requested
-    */}
-    <h1>Visits</h1>
-  </>)
-}
-
-export default VisitsView
